+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 # saltstack-age change log
+
+## Unreleased
+
+* fix: add support for nested pillar data